Add files via upload
Skeleton website pages with consistent nav bar. The first page is the 'About' page which gives the user a deeper understanding about why the website exists, an overview of what it contains, and a clear explanation of what the sustainable development goal chosen means to the authors and potentially the user. It could also include a mission statement of what the entity is trying to achieve. The 'Find Out More' page allows the user to access relevant articles or websites to support the activity of the entity.